Moves accordion to wrapper file structure
