+++ src/index.js
@@ -1,7 +1,7 @@
 // @flow
 
 // eslint-disable-next-line
-import Accordion from './Accordion/accordion';
+import Accordion from './Accordion/accordion-wrapper';
 import AccordionItem from './AccordionItem/accordion-item-wrapper';
 import AccordionItemTitle from './AccordionItemTitle/accordion-item-title';
 import AccordionItemBody from './AccordionItemBody/accordion-item-body';
